refactor: 렌주룰 Adapter적용
